brains/Math2matics2.brain/outputs/code-creation/agent_1766554350955_6fz075d/requirements.txt
# Dependencies required to run: python scripts/run_pipeline.py
# Keep versions minimally specified unless pinning is necessary.
__future__
ast
contextlib
difflib
entrypoint
fnmatch
importlib
io
jsonschema
matplotlib
numpy
Pillow
pipeline
platform
pytest
repro_json_pipeline
results_schema
run
secrets
shlex
socket
tensorflow
torch
toy_experiment
unicodedata
utils
